backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
cohere>=4.0.0
qdrant-client>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv>=0.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.25.0
asyncio>=3.4.0
aiofiles>=23.0.0
tiktoken>=0.5.0
pytest>=7.0.0
pytest-asyncio>=0.20.0